sequence logo plots now working

+++ src/utils/logomaker.py
@@ -1,49 +1,77 @@
-# standard imports
+import warnings
+
+warnings.filterwarnings("ignore")
 import io
+import os
 import base64
 import tempfile
-import numpy as np
-import pandas as pd
 import matplotlib.pyplot as plt
 
 import logomaker as lm
 
-from Bio import AlignIO, SeqIO
 from Bio.Align.Applications import ClustalwCommandline
-from io import StringIO
 
 
-def make_logo(seqs):  # Create a temporary file to hold the input sequences
-    temp_file = tempfile.NamedTemporaryFile(suffix=".fa", delete=False)
-    with open(temp_file.name, "w") as file:
+def are_all_strings_same_length(strings):
+    return len(set(len(s) for s in strings)) == 1
+
+
+def make_logo(seqs):
+    temp_in_file = tempfile.NamedTemporaryFile(suffix=".fa", delete=False)
+    temp_out_file = tempfile.NamedTemporaryFile(suffix=".fa", delete=False)
+
+    # Write sequences to the temporary input file
+    with open(temp_in_file.name, "w") as file:
         for idx, seq in enumerate(seqs):
             header = f">seq{idx + 1}"
             file.write(f"{header}\n{seq}\n")
 
-    clustalw_cline = ClustalwCommandline("clustalw", infile=temp_file.name)
-    clustalw_cline()
-    print(temp_file.name)
+    if os.path.exists(temp_in_file.name) and os.path.getsize(temp_in_file.name) > 0:
 
-    aligner = AlignIO.read(temp_file.name, "fasta")
+        # Run ClustalW for alignment
+        clustalw_cline = ClustalwCommandline(
+            "clustalw",
+            infile=temp_in_file.name,
+            outfile=temp_out_file.name,
+            output="FASTA",
+        )
+        clustalw_cline()
 
-    aligned_sequences = []
-    for record in aligner:
-        aligned_sequences.append(record.seq)
+        # Read the aligned sequences
+        with open(temp_out_file.name, "r") as f:
+            lines = f.readlines()
 
-    fig, ax = plt.subplots()
-    counts_mat = lm.alignment_to_matrix(
-        sequences=aligned_sequences, to_type="counts", characters_to_ignore=".-X"
-    )
+        aln_seqs = [
+            seq.strip().upper()
+            for seq in lines
+            if not seq.startswith("#") and not seq.startswith(">")
+        ]
 
-    lm.Logo(counts_mat, ax=ax)
+        # Check if all sequences are the same length
+        if are_all_strings_same_length(aln_seqs):
+            fig, ax = plt.subplots()
+            counts_mat = lm.alignment_to_matrix(
+                sequences=aln_seqs, to_type="counts", characters_to_ignore=".-X"
+            )
+            lm.Logo(counts_mat, ax=ax)
 
-    # Save the figure to a BytesIO object
-    buf = io.BytesIO()
-    fig.savefig(buf, format="png")
-    buf.seek(0)
+            # Save the figure to a BytesIO object
+            buf = io.BytesIO()
+            fig.savefig(buf, format="png")
+            buf.seek(0)
 
-    # Encode the image to base64
-    img_base64 = base64.b64encode(buf.getvalue()).decode("utf-8")
+            # Encode the image to base64
+            img_base64 = base64.b64encode(buf.getvalue()).decode("utf-8")
+            plt.close(fig)  # Close the figure to free up resources
 
-    plt.close(fig)  # Close the figure to free up resources
-    return img_base64
+            # Clean up temporary files
+            os.remove(temp_in_file.name)
+            os.remove(temp_out_file.name)
+            return img_base64
+        else:
+            # Clean up temporary files
+            os.remove(temp_in_file.name)
+            os.remove(temp_out_file.name)
+            return None
+    else:
+        return None
